Launch button + handler and game over win / lose screens added.

+++ VikingMathApp/VMAVikingPoolManager.h
@@ -24,6 +24,6 @@
 -(void)removeVikingFromPool;
 -(void)advanceVikingToOnPoint;
 -(void)layoutVikings:(BOOL)randomise;
--(int)numVikingsInPool;
+-(NSUInteger)numVikingsInPool;
 
 @end
